Safari-609.1.11.1/LayoutTests/fast/css/css2-system-fonts.html

<!DOCTYPE html>
This tests platform specific system font styles. If any of the styles appear in monospace the test fails.
<hr>
<div style="font-family: monospace;" id="container"></div>
<script>
var systemFontStyles = [
  'caption',
  'icon',
  'menu',
  'message-box',
  'small-caption',
  'status-bar',
  '-webkit-mini-control',
  '-webkit-small-control',
  '-webkit-control',
];
systemFontStyles.forEach(function(systemFont) {
  var item = document.createElement('div');
  container.appendChild(item);
  item.style.font = systemFont;
  item.textContent = systemFont + ': ' + getComputedStyle(item).font;
});
if (window.testRunner) {
  testRunner.dumpAsText();
}
</script>
