Safari-609.1.11.1/LayoutTests/platform/wpe/fast/css/css2-system-fonts-expected.txt

This tests platform specific system font styles. If any of the styles appear in monospace the test fails.
caption: normal normal normal normal 13px/normal monospace
icon: normal normal normal normal 13px/normal monospace
menu: normal normal normal normal 13px/normal monospace
message-box: normal normal normal normal 13px/normal monospace
small-caption: normal normal normal normal 13px/normal monospace
status-bar: normal normal normal normal 13px/normal monospace
-webkit-mini-control: normal normal normal normal 13px/normal monospace
-webkit-small-control: normal normal normal normal 13px/normal monospace
-webkit-control: normal normal normal normal 13px/normal monospace
